requirements.txt
aiogram==3.0.0b3
PyYAML==6.0
